Refactor UTMTrackingListener to conditionally delete messages and send replies

+++ src/listeners/utmtracking.ts
@@ -1,6 +1,7 @@
 import { Listener } from '@sapphire/framework';
 import { Message } from 'discord.js';
 
+const deleteMessage = false;
 export class UTMTrackingListener extends Listener {
 	constructor(context: Listener.LoaderContext, options: Listener.Options) {
 		super(context, {
@@ -12,7 +13,7 @@ export class UTMTrackingListener extends Listener {
 	public async run(message: Message) {
 		if (message.author.bot) return;
 		const urlRegex =
-			/(http|https):\/\/([\w_-]+(?:(?:\.[\w_-]+)+))([\w.,@?^=%&:/~+#-]*[\w@?^=%&\/~+#-])/g;
+			/(http|https):\/\/([\w_-]+(?:(?:\.[\w_-]+)+))([\w.,@?^=%&:/~+#-]*[\w@?^=%&/~+#-])/g;
 
 		const matches = Array.from(message.content.matchAll(urlRegex)).map(m => m[0]);
 		if (!matches || matches.length === 0) return;
@@ -23,13 +24,19 @@ export class UTMTrackingListener extends Listener {
 			modifiedMessage = modifiedMessage.replace(url, urlWithoutQueryParams);
 		}
 
-		// Reply emphemerally
-		await message.author.createDM(true).then(dmChannel => {
-			dmChannel.send(
-				`Your message has been modified to remove UTM parameters: \`\`\`${modifiedMessage}\`\`\``
+		if (deleteMessage) {
+			await Promise.allSettled([
+				message.delete(),
+				message.author.createDM(true).then(dmChannel => {
+					dmChannel.send(
+						`Your message has been modified to remove UTM parameters: \`\`\`${modifiedMessage}\`\`\``
+					);
+				}),
+			]);
+		} else {
+			await message.reply(
+				`Your message has been modified to remove UTM parameters: \n> ${modifiedMessage.replaceAll('\n', '\n> ')}`
 			);
-		});
-
-		await message.delete();
+		}
 	}
 }
